# Optimised code using arrays by avoiding LINQ and using Buffer.Copy