Performance characteristics of using Select(...).ToArray() vs. for loop

```
[Benchmark]
public string[] LinqSelectToArray_Current()
{
    string[] result = _bytes.Select(b => BitConverter.ToString(new byte[] {b})).ToArray();
    return result;
}

[Benchmark]
public string[] LinqSelectToArray_Proposed()
{
    string[] result = new string[_bytes.Length];

    for (int i = 0; i < _bytes.Length; i++)
    {
        result[i] = BitConverter.ToString(new byte[] {_bytes[i]});
    }

    return result;
}
```

|                     Method | Array Size|     Mean |     Error |    StdDev | Ratio | Rank | Gen 0/1k Op | Gen 1/1k Op | Gen 2/1k Op | Allocated Memory/Op |
|--------------------------- |-------------- |---------:|----------:|----------:|------:|-----:|------------:|------------:|------------:|--------------------:|
|  **LinqSelectToArray_Current** |             **32 elements** | **749.1 ns** |  **8.426 ns** |  **7.469 ns** |  **1.00** |    **1** |      **0.5655** |           **-** |           **-** |              **2376 B** |
|                            |               |          |           |           |       |      |             |             |             |                     |
| LinqSelectToArray_Proposed |             32 elements | 642.3 ns | 12.619 ns | 14.532 ns |  1.00 |    1 |      0.5541 |           - |           - |              2328 B |
|                            |               |          |           |           |       |      |             |             |             |                     |
|  **LinqSelectToArray_Current** |             **6 elements** | **200.8 ns** |  **3.101 ns** |  **2.901 ns** |  **1.00** |    **1** |      **0.1202** |           **-** |           **-** |               **504 B** |
|                            |               |          |           |           |       |      |             |             |             |                     |
| LinqSelectToArray_Proposed |             6 elements | 126.6 ns |  1.654 ns |  1.547 ns |  1.00 |    1 |      0.1087 |           - |           - |               456 B |

Using `ToArray()` as cloning mechanism

|            Method | Array Size |      Mean |     Error |    StdDev | Ratio | Rank | Gen 0/1k Op | Gen 1/1k Op | Gen 2/1k Op | Allocated Memory/Op |
|------------------ |-------------- |----------:|----------:|----------:|------:|-----:|------------:|------------:|------------:|--------------------:|
| **CopyArray_Current** |             **32 elements** | **51.687 ns** | **0.3557 ns** | **0.2970 ns** |  **1.00** |    **1** |      **0.0133** |           **-** |           **-** |                **56 B** |
|                   |               |           |           |           |       |      |             |             |             |                     |
| ToUInt32_Proposed |             32 elements |  9.848 ns | 0.1506 ns | 0.1409 ns |  1.00 |    1 |      0.0133 |           - |           - |                56 B |
|                   |               |           |           |           |       |      |             |             |             |                     |
| **CopyArray_Current** |             **6 elements** | **50.265 ns** | **0.2128 ns** | **0.1886 ns** |  **1.00** |    **1** |      **0.0076** |           **-** |           **-** |                **32 B** |
|                   |               |           |           |           |       |      |             |             |             |                     |
| ToUInt32_Proposed |             6 elements |  8.543 ns | 0.0939 ns | 0.0878 ns |  1.00 |    1 |      0.0076 |           - |           - |                32 B |