@MithrilMan 2019-03-13
as soon as we'll move to target dotnetcore2.1 (well actually 2.2) on every project, we'll have fun about optimizing memory access with Span/ReadOnlySpan and ArrayPool<T>, I wouldn't spend much time optimizing now