@maciejzaleski 2019-03-14
I went for the low hanging fruits here. Things like:
- Calling array.Reverse().ToArray() where we have methods which do not require reversing
- Using ToArray() as a cloning mechanism for arrays is 5 times slower than Array.Copy
- Using List where we know the number of elements is probably tiny bit less efficient than using an array
- Using LINQ over foreach loop is marginally slower