@dangershony 2019-03-15
There is definitely a benefit in moving to `Array.Copy` how ever this PR has more then just that.